Improve nodes querying
We've got a few similar stacktraces once. See the following one for
example:

** Reason for termination ==
** {badarg,
       [{ets,next,[sys_dist,'rabbitmq-cli-42@host.example.com'],[]},
        {net_kernel,get_nodes,2,[{file,"net_kernel.erl"},{line,1025}]},
        {net_kernel,get_nodes,2,[{file,"net_kernel.erl"},{line,1019}]},
        {net_kernel,get_nodes_info,0,[{file,"net_kernel.erl"},{line,1439}]},
        {rabbit_mgmt_external_stats,cluster_links,0,
            [{file,"src/rabbit_mgmt_external_stats.erl"},{line,252}]},
        {rabbit_mgmt_external_stats,emit_node_node_stats,1,
            [{file,"src/rabbit_mgmt_external_stats.erl"},{line,366}]},
        {rabbit_mgmt_external_stats,handle_info,2,
            [{file,"src/rabbit_mgmt_external_stats.erl"},{line,347}]},
        {gen_server,try_dispatch,4,[{file,"gen_server.erl"},{line,615}]}]}

The problem is that when we're trying to query a list of connected
nodes, we're doing it in the following way:

  Call for the first record in ETS
  While not EOF:
    Call for the next record in ETS

What happens, when some Node disconnects during the "not EOF" loop?
We'll get an exception.

Let's do it differently - query a list of nodes in one shot, and then
get info from each of the nodes in list (w/o extra calls to ets). These
individual calls care of disconnected nodes so everything will be fine
even if a node disconnects.

Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>