erl_child_setup: reduce number of calls to close()
On systems without closefrom(), such as Linux, iterating over all
possible file descriptors and calling close() for each is inefficient.
This is markedly so when the maximum number of file descriptors has
been tuned to a large number.

Instead, walk the open descriptors under /dev/fd and close only those
which are open.